Ensure Cargo.lock is up-to-date
`Cargo.lock` must always be up-to-date if this was a binary crate,
because users may try to install it via `cargo install` instead of
`cargo install --locked`.

`lazy_errors` is not a binary crate. Cargo does not publish `Cargo.lock`
to crates.io. When users add a dependency on `lazy_errors`, Cargo will
try to use recent versions of (transitive) dependencies, regardless
of what we put in our `Cargo.lock`. Cargo may also fall back to older
semver-compatible versions, depending on requirements of other
dependencies of the user. Thus, `lazy_errors` must support
using recent dependency versions (i.e. after running `cargo update`),
but having an outdated `Cargo.lock` doesn't matter to the user,
so we have allowed it to be outdated so far.

However, if `Cargo.lock` is outdated and the software is only tested
after `cargo update` in CI, we may introduce features that actually
don't work in local checkouts without updating `Cargo.lock` before.
Having to run `cargo update` locally must not be a precondition
to be able to build/test this software.Thus, we have two options:
Either run all steps of the CI pipeline both with and without updating
`Cargo.lock` before, effectively doubling CI time/resource usage.
Or simply treat binary and library projects the same and require
`Cargo.lock` to be up-to-date in all cases.

Requiring `Cargo.lock` to be up-to-date will make the (scheduled) build
fail more often. However, this can be fixed very quickly and does not
mandate publishing a release of `lazy_errors` because it's a library.