Add StashedResult::ok()
Returns `Some(t)` if `self` is `Ok(t)`, `None` otherwise.

This method is useful to discard the `&mut` borrowing of the
`ErrorStash`/`StashWithErrors` that was passed as parameter
to `or_stash`. You may need to do this if you have multiple
`or_stash` statements and want to extract the `Ok(T)` result
from them later.