Skip validating moving targets in publish.yaml
We're only doing a quick validation in publish.yaml.
Calling `validate.yaml` introduces a “race condition”
with the rest of the Rust ecosystem: `validate.yaml`
checks several “moving targets”, such as whether
`Cargo.lock` is up-to-date.

Thus, any commit that had already passed the `validate.yaml` check
may fail it if it is checked again at some later point in time,
for example when a new version of any (transitive) dependency
gets released. Since we've validated each commit on the main branch,
all commits on the main branch are good to release per se.
If we rejected such a release just because we're a few minutes
“too late” (whatever that means), we would have to publish
that version manually or we would have to delete the git tag.
Even if we noticed before creating the tag, this would require
us to create a “hotfix” branch for completely artificial reasons.