Deactivate automatic GitHub Release generation
Instead, create the releases manually for now by pasting
the relevant parts from `CHANGELOG.md` into the GitHub Web UI.
It's not much of an effort and the release page will then contain
text that's sized appropriately and not monospaced.

When the changelog generation process is automated, use the
_new_ part of the changelog as parameter for `body` or `body_path`
of the `softprops/action-gh-release@v2` GitHub Action.